Data component changes
Added IDs in App state that can be changed to search different data from
backend by utilizing the functions in the Data component (getRegionLevels,
getRegions and getScenarioCollections).
Added comments to Data.js on how the functions work.
Added <p> tags to split the data holders from each other (these will
display the IDs that are set in the app state.